+++ src/app/admin/companies/companies.component.ts
@@ -6,6 +6,10 @@ import { Breadcrumb } from "app/layout/components/content-header/breadcrumb/brea
 import { adminMenu } from "app/menu/adminMenu";
 import { Company } from "./company";
 import { CompanyService } from "./company.service";
+import {NgbModal} from '@ng-bootstrap/ng-bootstrap';
+import {User} from '../../utils/common/login/user';
+import {UserServiceService} from '../../utils/common/login/user-service.service';
+import {Router} from '@angular/router';
 
 @Component({
   selector: "app-companies",
@@ -19,12 +23,17 @@ export class CompaniesComponent implements OnInit {
   coreConfig: any;
   menu: any;
   listCompanies: Company[] = [];
+  newSociete = new Company ();
+  modifSociete = new Company ();
 
   constructor(
     private _coreSidebarService: CoreSidebarService,
     private _coreMenuService: CoreMenuService,
     private _coreConfigService: CoreConfigService,
-    private companyService: CompanyService
+    private companyService: CompanyService,
+    private modalService: NgbModal,
+    private userService: UserServiceService,
+    private _router: Router,
   ) {
     this.menu = adminMenu;
     this._coreConfigService.setConfig({
@@ -70,27 +79,50 @@ export class CompaniesComponent implements OnInit {
     };
   }
 
+
+  modalOpen(modalForm) {
+    console.log(modalForm);
+    this.modalService.open(modalForm);
+  }
+
+   goToModal(a: any) {
+    this.modifSociete = a;
+  }
+
   async onGetCompanies() {
     this.listCompanies = await this.companyService.getCompanies();
+  this.listCompanies.forEach(async a => {
+    a.listClient = await this.userService.getClientsByCompany(a.id);
+  })
   }
 
   async onSaveCompant() {
-    await this.companyService.saveCompany(this.company);
+    await this.companyService.saveCompany(this.newSociete);
+    this.ngOnInit();
   }
 
   async onEditCompany() {
-    await this.companyService.editCompany(this.company);
+    await this.companyService.editCompany(this.modifSociete);
+    this.ngOnInit();
   }
 
   async onDeleteCompany(id: number) {
-    await this.companyService.deleteCompany(id);
+    if(confirm("Are you sure to delete this company")) {
+      await this.companyService.deleteCompany(id);
+    this.ngOnInit();
+    }
+
   }
 
   GetInitials(name) {
     return name.charAt(0).toUpperCase();
   }
 
-  ngOnInit(): void {}
+  async ngOnInit() {
+    if (localStorage.getItem('connected') == 'no')
+      this._router.navigate(["/login"]);
+    await this.onGetCompanies();
+  }
 
   /**
    * Toggle the sidebar
@@ -100,4 +132,6 @@ export class CompaniesComponent implements OnInit {
   toggleSidebar(key): void {
     this._coreSidebarService.getSidebarRegistry(key).toggleOpen();
   }
+
+
 }
